more rpcendpoint

+++ internal/infra/cluster/service/rpcservice/dto/rpcresponse.go
@@ -1,7 +1,7 @@
 package dto
 
 type RpcResponse[T comparable] struct {
-	RequestId int
+	RequestId int64
 	Result    T
 	Rpcerror  error
 }
